ee_map update
A standard Generic + (S4) is now implemented in rgee to make possible adds data from the second map to the first one just applying the sum operator (e.g. m1 + m2). It makes that a token return function does not been necessary. This implementation was based on the codes of tim-salabim .